requirements.txt
-e git+https://algonauts@dev.azure.com/algonauts/algonautsutils/_git/algonautsutils@6835d48f3718bac9a57d5d700216f5760ba40db4#egg=algonautsutils
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
Babel==2.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-crontab==0.7.1
django-extensions==2.2.5
django-phonenumber-field==4.0.0
django-secure==1.0.1
django-sslserver==0.22
django-unixtimestampfield==0.3.9
django-webpush==0.3.1
enum34==1.1.9
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
http-ece==1.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiteconnect==3.7.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
phonenumbers==8.11.2
psycopg2==2.8.4
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.9.4
pywin32==227
razorpay==1.2.0
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
urllib3==1.25.7
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.7.1
